refactor: useEmailCheck 훅 만들어 비즈니스 로직 은닉
